acs-decision/content-moderator/text-moderation/requirements.txt
azure-cognitiveservices-vision-contentmoderator
python-dotenv
